Add files via upload
This contains the project files for project 4 from CS 559. The requirements were to create a program that uses a parametric curve to define an intricate trajectory, and use this representation both to draw such a curve as well as move/animate an object along it. I went above and beyond, however, by doing all of this in 3D.